pi-server/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pyfingerprint>=2.0.0
mfrc522>=1.0.0
RPi.GPIO>=0.7.0
pyserial>=3.5
